+++ .github/workflows/ci.yml
@@ -1,4 +1,4 @@
-name: Tests and Coverage
+name: Python Quality
 
 on:
   push:
@@ -9,8 +9,30 @@ permissions:
   id-token: write
 
 jobs:
+  lint:
+    name: Lint
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
+          cache: pip
+
+      - name: Install package
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install -e ".[dev]"
+
+      - name: Run lint
+        run: python -m ruff check .
+
   tests:
-    name: Python 3.13
+    name: Tests and Coverage
     runs-on: ubuntu-latest
 
     steps:
