+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
 dev = [
   "pytest>=8.0",
   "pytest-cov>=5.0",
+  "ruff>=0.8",
   "syrupy>=4.6",
 ]
 
@@ -33,3 +34,6 @@ packages = ["src/bidsflow"]
 
 [tool.pytest.ini_options]
 addopts = "-q"
+
+[tool.ruff]
+target-version = "py311"
