+++ src/bidsflow/heudiconv/__init__.py
@@ -1,0 +1,70 @@
+from __future__ import annotations
+
+from .common import (
+    HeudiconvInitError,
+    HeudiconvRunError,
+    format_command,
+)
+from .draft import (
+    DraftPlan,
+    DraftResult,
+    DraftUnitPlan,
+    DraftUnitResult,
+    HeudiconvDraftError,
+    plan_draft,
+    run_draft,
+)
+from .init import InitPlan, InitResult, plan_heudiconv_init, run_heudiconv_init
+from .run import (
+    RunPlan,
+    RunResult,
+    RunUnitPlan,
+    RunUnitResult,
+    SgeRunPlan,
+    plan_heudiconv_run,
+    preview_run_unit_selection,
+    run_heudiconv,
+)
+from .sources import (
+    SourcesEntry,
+    SourcesError,
+    SourcesPlan,
+    SourcesResult,
+    list_sources_review_issues,
+    plan_sources,
+    run_sources,
+    summarize_sources_entries,
+)
+
+__all__ = [
+    "DraftPlan",
+    "DraftResult",
+    "DraftUnitPlan",
+    "DraftUnitResult",
+    "HeudiconvDraftError",
+    "HeudiconvInitError",
+    "HeudiconvRunError",
+    "InitPlan",
+    "InitResult",
+    "RunPlan",
+    "RunResult",
+    "RunUnitPlan",
+    "RunUnitResult",
+    "SgeRunPlan",
+    "SourcesEntry",
+    "SourcesError",
+    "SourcesPlan",
+    "SourcesResult",
+    "format_command",
+    "list_sources_review_issues",
+    "plan_draft",
+    "plan_heudiconv_init",
+    "plan_heudiconv_run",
+    "plan_sources",
+    "preview_run_unit_selection",
+    "run_draft",
+    "run_heudiconv",
+    "run_heudiconv_init",
+    "run_sources",
+    "summarize_sources_entries",
+]
