+++ tests/test_heudiconv_helper_edges.py
@@ -6,6 +6,10 @@ from pathlib import Path
 import pytest
 
 import bidsflow.heudiconv as h
+import bidsflow.heudiconv.common as heudiconv_common
+import bidsflow.heudiconv.draft as heudiconv_draft
+import bidsflow.heudiconv.run as heudiconv_run
+import bidsflow.heudiconv.sources as heudiconv_sources
 from helpers import read_key_value_file
 
 
@@ -24,8 +28,8 @@ def test_low_level_run_helpers_cover_status_paths(tmp_path: Path, monkeypatch) -
         claim_path=tmp_path / "claims" / "sub-001.running",
     )
 
-    assert not h._unit_status_is_succeeded(unit.status_path)
-    h._write_unit_status(
+    assert not heudiconv_run._unit_status_is_succeeded(unit.status_path)
+    heudiconv_run._write_unit_status(
         unit,
         status="succeeded",
         backend="sge",
@@ -38,7 +42,7 @@ def test_low_level_run_helpers_cover_status_paths(tmp_path: Path, monkeypatch) -
         scheduler_task_id=None,
         error="multi\nline\rerror",
     )
-    assert h._unit_status_is_succeeded(unit.status_path)
+    assert heudiconv_run._unit_status_is_succeeded(unit.status_path)
     payload = read_key_value_file(unit.status_path)
     assert payload["task_id"] == ""
     assert payload["error"] == "multi line error"
@@ -51,11 +55,11 @@ def test_low_level_run_helpers_cover_status_paths(tmp_path: Path, monkeypatch) -
         return original_read_text(path, *args, **kwargs)
 
     monkeypatch.setattr(Path, "read_text", _raise_os_error)
-    assert not h._unit_status_is_succeeded(unit.status_path)
+    assert not heudiconv_run._unit_status_is_succeeded(unit.status_path)
     monkeypatch.undo()
 
-    h._release_unit_claim(unit)
-    h._write_unit_claim(unit=unit)
+    heudiconv_run._release_unit_claim(unit)
+    heudiconv_run._write_unit_claim(unit=unit)
     assert unit.claim_path.exists()
     other_unit = replace(
         unit,
@@ -63,13 +67,13 @@ def test_low_level_run_helpers_cover_status_paths(tmp_path: Path, monkeypatch) -
         claim_path=tmp_path / "claims" / "sub-002.running",
         status_path=tmp_path / "state" / "sub-002.status",
     )
-    h._write_unit_claim(unit=other_unit)
-    h._release_unfinished_claims((unit, other_unit), [], unit)
+    heudiconv_run._write_unit_claim(unit=other_unit)
+    heudiconv_run._release_unfinished_claims((unit, other_unit), [], unit)
     assert not unit.claim_path.exists()
     assert not other_unit.claim_path.exists()
 
     unit.status_path.write_text("# comment\nignored\nstatus=succeeded\n", encoding="utf-8")
-    assert h._read_key_value_status(unit.status_path) == {"status": "succeeded"}
+    assert heudiconv_run._read_key_value_status(unit.status_path) == {"status": "succeeded"}
 
 
 def test_low_level_path_and_scheduler_helpers(tmp_path: Path) -> None:
@@ -78,20 +82,20 @@ def test_low_level_path_and_scheduler_helpers(tmp_path: Path) -> None:
     outside_path = tmp_path / "outside"
     outside_path.mkdir()
 
-    assert not h._cleanup_run_execution_view(project_root, outside_path)
+    assert not heudiconv_run._cleanup_run_execution_view(project_root, outside_path)
     with pytest.raises(ValueError, match="outside the project root"):
-        h._remove_project_path(project_root, outside_path)
+        heudiconv_common._remove_project_path(project_root, outside_path)
     with pytest.raises(h.HeudiconvRunError, match="must name an immediate child directory"):
-        h._resolve_sources_source_path(project_root, "../SUB001")
+        heudiconv_sources._resolve_sources_source_path(project_root, "../SUB001")
 
     log_path = project_root / "logs" / "unit.log"
-    h._append_log(log_path, "first")
-    h._append_log(log_path, "second")
+    heudiconv_common._append_log(log_path, "first")
+    heudiconv_common._append_log(log_path, "second")
     assert log_path.read_text(encoding="utf-8") == "first\n\nsecond\n"
 
-    assert h._combine_process_output("out\n", "err\n") == "out\nerr"
-    assert h._parse_sge_job_id(None) == ""
-    assert h._parse_sge_job_id("123.1 first line\nignored") == "123.1 first line"
+    assert heudiconv_common._combine_process_output("out\n", "err\n") == "out\nerr"
+    assert heudiconv_run._parse_sge_job_id(None) == ""
+    assert heudiconv_run._parse_sge_job_id("123.1 first line\nignored") == "123.1 first line"
 
     sources_plan = h.SourcesPlan(
         source_root=project_root / "sourcedata",
@@ -104,11 +108,11 @@ def test_low_level_path_and_scheduler_helpers(tmp_path: Path) -> None:
     sources_plan.sources_path.parent.mkdir(parents=True)
     sources_plan.sources_path.write_text("existing\n", encoding="utf-8")
     with pytest.raises(h.SourcesError, match="Existing BIDSFlow sources state"):
-        h._guard_sources_reset_requirement(sources_plan, reset=False)
+        heudiconv_sources._guard_sources_reset_requirement(sources_plan, reset=False)
     outside_sources_plan = replace(sources_plan, sources_path=tmp_path / "outside.tsv")
     (tmp_path / "outside.tsv").write_text("outside\n", encoding="utf-8")
     with pytest.raises(h.SourcesError, match="outside the project root"):
-        h._reset_sources_state(project_root, outside_sources_plan)
+        heudiconv_sources._reset_sources_state(project_root, outside_sources_plan)
 
     draft_plan = h.DraftPlan(
         sample_paths=(),
@@ -124,14 +128,14 @@ def test_low_level_path_and_scheduler_helpers(tmp_path: Path) -> None:
     )
     draft_plan.draft_work_root.mkdir()
     with pytest.raises(h.HeudiconvDraftError, match="outside the project root"):
-        h._reset_draft_state(project_root, draft_plan)
+        heudiconv_draft._reset_draft_state(project_root, draft_plan)
 
 
 def test_make_executable_sets_posix_execute_bit(tmp_path: Path) -> None:
     executable = tmp_path / "script.sh"
     executable.write_text("#!/bin/sh\n", encoding="utf-8", newline="\n")
 
-    h._make_executable(executable)
+    heudiconv_common._make_executable(executable)
 
     assert executable.stat().st_mode & 0o111
 
@@ -147,7 +151,7 @@ def test_materialize_run_execution_view_error_paths(tmp_path: Path, monkeypatch)
         symlink_calls.append((Path(target), target_is_directory))
 
     monkeypatch.setattr(Path, "symlink_to", _record_symlink)
-    h._materialize_run_execution_view(execution_path, source_path)
+    heudiconv_run._materialize_run_execution_view(execution_path, source_path)
     assert symlink_calls == [(source_path, True)]
 
     def _raise_symlink_error(self, target, target_is_directory=False):
@@ -155,4 +159,4 @@ def test_materialize_run_execution_view_error_paths(tmp_path: Path, monkeypatch)
 
     monkeypatch.setattr(Path, "symlink_to", _raise_symlink_error)
     with pytest.raises(h.HeudiconvRunError, match="Failed to create temporary execution link"):
-        h._materialize_run_execution_view(execution_path, source_path)
+        heudiconv_run._materialize_run_execution_view(execution_path, source_path)
